Add refactored 'teacher/apprentice_TEACHER' classes
